requirements.txt
python-dotenv
gitpython
chevron
colorama